+++ CHANGELOG.md
@@ -7,9 +7,9 @@
 ## New Feature
 - \#312 - Add doc badge
 - \#319 - Support auto merge PR after reviewed
+- \#333, \#346 - Change method insert() to bulk_insert()
 - \#344 - Add pylint check
 - \#345 - Do not support python 3.5.x any more
-- \#346 - Change method insert() to bulk_insert()
 
 ## Improvement
 - \#277 - Add doc `FAQ` chapter of pymilvus documentation